Revert "More debug"
This reverts commit 3c5c29783cdb7fb954d88e13473f4d8d32cd7fb3.

+++ Assets/Scripts/UI/UIManager.cs
@@ -45,7 +45,7 @@ public class UIManager : MonoBehaviour
 
     public static UIManager instance;
 
-    public  static int debugTimeStamp;
+    public  static long debugTimeStamp;
     [SerializeField]
     private Text debugText;
     public static void StartGame()
